ai-service/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
pymupdf
python-dotenv
pydantic
pydantic-settings
httpx
python-multipart
langchain-text-splitters
numpy
schedule
cloudinary
langchain-groq
